add user

+++ api/repository/pictures.py
@@ -3,13 +3,13 @@ from typing import List, Type
 from sqlalchemy.orm import Session
 
 from api.conf.config import settings
-from api.database.models import Picture, Tag
+from api.database.models import Picture, Tag, User
 from api.repository.tags import create_tag
 from api.schemas import PictureCreate
 from api.services.cloud_picture import CloudImage
 
 
-async def create_picture(description: str, tags: List[str], file_path: str, db: Session):
+async def create_picture(description: str, tags: List[str], file_path: str, db: Session, user: User):
     # If the number of tags is greater than the maximum, return an error message
     if len(tags) > settings.max_tags:
         return f"Error: Too many tags. The maximum is {settings.max_tags}."
@@ -18,11 +18,11 @@ async def create_picture(description: str, tags: List[str], file_path: str, db: 
     if tags:
         tags_list = await transformation_list_to_tag(tags[0].split(","), db)
 
-    picture = Picture(picture_url=file_path, description=description, tags=tags_list)
+    picture = Picture(picture_url=file_path, description=description, tags=tags_list, user_id=user.id)
     db.add(picture)
     db.commit()
     db.refresh(picture)
-    # , user_id = user.id
+
     return picture
 
 
@@ -32,12 +32,11 @@ def get_tag_by_name(tag_name: str, db: Session) -> Tag | None:
 
 
 async def transformation_list_to_tag(tags: list, db: Session) -> List[Tag]:
-    # , user
+
     list_tags = []
     if tags:
         for tag_name in tags:
             tag = await create_tag(tag_name, db)
-            # user,
             list_tags.append(tag)
     return list_tags
 
@@ -52,28 +51,35 @@ async def get_user_pictures(user_id: int, db: Session) -> list[Type[Picture]]:
     return pictures
 
 
-async def remove_picture(picture_id: int, db: Session):
+async def remove_picture(picture_id: int, user: User, db: Session):
 
     picture = db.query(Picture).filter(Picture.id == picture_id).first()
     if picture:
-        public_id = picture.picture_url.split("/")[-1]
-        CloudImage.destroy(public_id)
-        db.delete(picture)
-        db.commit()
+
+        # TODO - кому дозволимо видаляти , наприклад - адміністратор може видаляти, що хоче, а юзер - свої
+        # if user.role == admin or picture.user_id == user.id
+        if picture.user_id == user.id:
+            public_id = picture.picture_url.split("/")[-1]
+            CloudImage.destroy(public_id)
+            db.delete(picture)
+            db.commit()
     return picture
 
 
-async def update_picture(picture_id: int, body: PictureCreate, db: Session):
-    # , user: User
+async def update_picture(picture_id: int, body: PictureCreate, user: User, db: Session):
+
     picture = db.query(Picture).filter(Picture.id == picture_id).first()
 
     if picture:
-        tags_list = transformation_list_to_tag(body.tags, db)
-        # user,
-        picture.description = body.description
-        picture.tags = tags_list
-        db.commit()
-        db.refresh(picture)
+
+        # TODO - кому дозволимо видаляти , наприклад - адмністратор може видаляти, що хоче, а юзер - свої
+        # if user.role == admin or picture.user_id == user.id
+        if picture.user_id == user.id:
+            tags_list = transformation_list_to_tag(body.tags, db)
+            picture.description = body.description
+            picture.tags = tags_list
+            db.commit()
+            db.refresh(picture)
     return picture
